Fix the raw single column distinct executors (#6411)
- Fix the typos within the raw single column distinct executor
- Add unit test for that

+++ pinot-core/src/main/java/org/apache/pinot/core/query/distinct/raw/BaseRawLongSingleColumnDistinctExecutor.java
@@ -50,7 +50,7 @@ abstract class BaseRawLongSingleColumnDistinctExecutor implements DistinctExecut
   @Override
   public DistinctTable getResult() {
     DataSchema dataSchema =
-        new DataSchema(new String[]{_expression.toString()}, new ColumnDataType[]{ColumnDataType.DOUBLE});
+        new DataSchema(new String[]{_expression.toString()}, new ColumnDataType[]{ColumnDataType.LONG});
     List<Record> records = new ArrayList<>(_valueSet.size());
     LongIterator valueIterator = _valueSet.iterator();
     while (valueIterator.hasNext()) {
